pyproject.toml
[project]
name = "mpn-rl"
version = "0.1.0"
description = "Multi-Plasticity Networks for Reinforcement Learning with PCA Analysis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "gymnasium>=0.29.0",
    "scikit-learn>=1.3.0",
    "imageio>=2.31.0",
    "pillow>=10.0.0",
    "tensordict>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
]
